Clean the code and rename some files
